apps/requirements.txt
apache-airflow
pandas
numpy
praw
python-dotenv
s3fs